Update named1.conf
there was an Err in Debian based Distros, when you want to write views there was a problem and Err that said when you use view statement you must add all zones in views , so in line 21 there was a zone. so in debian based distros you didnt saw that zone.